update wifi script

+++ root/RucoPie/scripts/wifi.sh
@@ -3,25 +3,33 @@
 ## Based on gist.github.com/rjsteinert
 ## usage:
 ## sudo ./wifi.sh <ssid> <pass>
+#!/bin/bash
 
-ifdown wlan0
+wget -q --spider http://google.com
 
-# build the interfaces file that will point to the file that holds our configuration
-rm /etc/network/interfaces
-touch /etc/network/interfaces
-echo "auto lo
-iface lo inet loopback
-iface eth0 inet dhcp
-allow-hotplug wlan0
-iface wlan0 inet manual
-wpa-roam /etc/wpa_supplicant/wpa_supplicant.conf
-iface default inet dhcp" >> /etc/network/interfaces
+if [ $? -eq 0 ]; then
+    echo "rucopie-already-online"
+else
+    ifdown wlan0
 
-# build the supplicant file that holds our configuration
-rm /etc/wpa_supplicant/wpa_supplicant.conf
-touch /etc/wpa_supplicant/wpa_supplicant.conf
-echo 'ctrl_interface=DIR=/var/run/wpa_supplicant GROUP=netdev' >> /etc/wpa_supplicant/wpa_supplicant.conf
-echo 'update_config=1' >> /etc/wpa_supplicant/wpa_supplicant.conf
+    # build the interfaces file that will point to the file that holds our configuration
+    rm /etc/network/interfaces
+    touch /etc/network/interfaces
+    echo "auto lo
+    iface lo inet loopback
+    iface eth0 inet dhcp
+    allow-hotplug wlan0
+    iface wlan0 inet manual
+    wpa-roam /etc/wpa_supplicant/wpa_supplicant.conf
+    iface default inet dhcp" >> /etc/network/interfaces
 
-wpa_passphrase "$1" "$2" >> /etc/wpa_supplicant/wpa_supplicant.conf
-ifup wlan0
+    # build the supplicant file that holds our configuration
+    rm /etc/wpa_supplicant/wpa_supplicant.conf
+    touch /etc/wpa_supplicant/wpa_supplicant.conf
+    echo 'ctrl_interface=DIR=/var/run/wpa_supplicant GROUP=netdev' >> /etc/wpa_supplicant/wpa_supplicant.conf
+    echo 'update_config=1' >> /etc/wpa_supplicant/wpa_supplicant.conf
+
+    wpa_passphrase "$1" "$2" >> /etc/wpa_supplicant/wpa_supplicant.conf
+    ifup wlan0
+    echo "rucopie-connected"
+fi
